organized simulationservice code
